split entities
