+++ shared/actions/entities.js
@@ -1,6 +1,6 @@
 // @flow
 import * as I from 'immutable'
-import type {Delete, Merge, Replace, Subtract, EntityType} from '../constants/entities'
+import type {Delete, Merge, Replace, Subtract, EntityType} from '../constants/types/entities'
 
 const deleteEntity: (keyPath: Array<string>, ids: I.List<string>) => Delete = (keyPath, ids) => ({
   payload: {ids, keyPath},
