+++ shared/constants/reducer.js
@@ -4,7 +4,7 @@ import type {State as Chat} from '../constants/types/chat'
 import type {State as Config} from '../constants/types/config'
 import type {State as Dev} from '../constants/types/dev'
 import type {State as Devices} from '../constants/types/devices'
-import type {State as Entity} from '../constants/entities'
+import type {State as Entity} from '../constants/types/entities'
 import type {State as Engine} from '../constants/engine'
 import type {State as Favorite} from '../constants/favorite'
 import type {State as Gregor} from '../constants/gregor'
